refactor: react import, consolidate components
