+++ packages/splits-kit/src/components/CreateSplit/ControllerSelector.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import {
   Control,
   UseFormSetValue,
