+++ packages/splits-kit/src/components/DisplaySplit/SkeletonLoader.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { range } from 'lodash'
 
 const SkeletonLoader = () => {
