+++ packages/splits-kit/src/components/inputs/AddressInput.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 import {
   Control,
   Controller,
