+++ packages/splits-kit/src/components/inputs/InputRow.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { InformationCircleIcon } from '@heroicons/react/20/solid'
 import Tooltip from '../util/Tooltip'
