+++ packages/splits-kit/src/components/inputs/NumberInput.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import {
   Control,
   Controller,
