+++ packages/splits-kit/src/components/util/DropdownMenu.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Menu } from '@headlessui/react'
 
 interface MenuItem {
