+++ packages/splits-kit/src/components/util/Label.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 export default function Label({
   icon,
   text,
