+++ packages/splits-kit/storybook/components/ConnectWallet.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import {
   useAccount,
   useConnect,
