+++ packages/splits-kit/storybook/preview.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import WagmiProvider from './components/WagmiProvider'
 import { dark, light } from './theme'
