+++ packages/splits-kit/storybook/stories/CreateSplit.stories.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import type { Meta, StoryObj } from '@storybook/react'
 import { CreateSplit } from '../../src'
 import ConnectWallet from '../components/ConnectWallet'
