+++ packages/splits-kit/storybook/stories/DisplaySplit.stories.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import type { Meta, StoryObj } from '@storybook/react'
 import { DisplaySplit } from '../../src'
 import ConnectWallet from '../components/ConnectWallet'
