+++ packages/splits-kit/storybook/stories/Introduction.stories.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { CreateSplit, DisplaySplit } from '../../src'
 import { SplitsProvider } from '@0xsplits/splits-sdk-react'
