+++ packages/splits-sdk-react/src/context.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useState, useMemo } from 'react'
+import { createContext, useState, useMemo } from 'react'
 import { SplitsClient, SplitsClientConfig } from '@0xsplits/splits-sdk'
 
 export type SplitsReactSdkContext = {
